Modify tests for Database engine
