+++ src/test/java/repackage/net/sf/persist/tests/engine/framework/BytesBlob.java
@@ -1,7 +1,7 @@
 
 // $Id$
 
-package repackage.net.sf.persist.tests.product.framework;
+package repackage.net.sf.persist.tests.engine.framework;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
