+++ src/test/java/repackage/net/sf/persist/tests/engine/framework/FieldMap.java
@@ -1,7 +1,7 @@
 
 // $Id$
 
-package repackage.net.sf.persist.tests.product.framework;
+package repackage.net.sf.persist.tests.engine.framework;
 
 import java.util.ArrayList;
 import java.util.List;
