+++ src/test/java/repackage/net/sf/persist/tests/engine/package-info.java
@@ -3,5 +3,5 @@
  *
  * @see {@link https://github.com/r5v9/persist/tree/master/src/tests/net/sf/persist/tests}
  */
-package repackage.net.sf.persist.tests.product;
+package repackage.net.sf.persist.tests.engine;
