+++ src/test/java/repackage/net/sf/persist/tests/engine/postgresql/schema.sql
@@ -34,5 +34,3 @@ create table binary_types (
 	blob_col oid
 );
 
-
-
