+++ src/test/java/repackage/net/sf/persist/tests/engine/sqlserver/schema.sql
@@ -1,16 +1,10 @@
-
-create table simple (
-	id int identity(1,1),
-	string_col varchar(255),
-	int_col int,
-	primary key (id)
-);
+drop table if exists numeric_types, datetime_types, string_types, binary_types;
 
 create table numeric_types (
-	bit_col bit,	
+	bit_col bit,
 	tinyint_col tinyint,
 	smallint_col smallint,
-	int_col int,	
+	int_col int,
 	bigint_col bigint,
 	smallmoney_col smallmoney,
 	money_col money,
@@ -21,17 +15,14 @@ create table numeric_types (
 );
 
 create table datetime_types (
-	datetime_col datetime,
-	smalldatetime_col smalldatetime
+	datetime_col datetime
 );
 
 create table string_types (
 	char_col char,
 	varchar_col varchar(255),
-	text_col text,
 	nchar_col nchar,
-	nvarchar_col nvarchar(255),
-	ntext_col ntext
+	nvarchar_col nvarchar(255)
 );
 
 create table binary_types (
@@ -39,3 +30,4 @@ create table binary_types (
 	varbinary_col varbinary(255),
 	image_col image
 );
+
